Revert "feat: Create Log - Info Tab"
