+++ README.md
@@ -45,5 +45,3 @@ If formatting doesn't work, download Prettier through `npx add
 ## Figma
 Designed by (TODO) 
 https://www.figma.com/file/wpNFTFqDPzdsiyElvKO5Qr/Main-App-Flow-%2F-Requirements?node-id=16%3A142&t=wEGgdD8OX8Ai83Dn-1
-
-## Design Docs
