--- docs/Design Docs/V0/Create Log.md
@@ -1,25 +1,0 @@
-# Overview
-- Status: 
-- Owner: 
-- Contributors: 
-- Goals: 
-
-# Table of Contents
-- [Overview](#overview)
-- [Table of Contents](#table-of-contents)
-- [Requirements/Goals](#requirementsgoals)
-- [Glossary](#glossary)
-- [Design Diagrams](#design-diagrams)
-- [Tech Decisions](#tech-decisions)
-  - [Taking Input From users](#taking-input-from-users)
-
-# Requirements/Goals
-
-# Glossary
-
-# Design Diagrams
-
-# Tech Decisions
-
-## Taking Input From users
-Options: React Hook Forms, FormiK/Yup, Manual data validation from scratch/ useStates
