--- docs/Design Docs/V0/Pottery Diary.md
@@ -1,21 +1,0 @@
-# Overview
-- Status: 
-- Owner: 
-- Contributors: 
-- Goals: 
-
-# Table of Contents
-- [Overview](#overview)
-- [Table of Contents](#table-of-contents)
-- [Requirements/Goals](#requirementsgoals)
-- [Glossary](#glossary)
-- [Design Diagrams](#design-diagrams)
-- [Tech Decisions](#tech-decisions)
-
-# Requirements/Goals
-
-# Glossary
-
-# Design Diagrams
-
-# Tech Decisions
