+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "@react-navigation/bottom-tabs": "^6.5.7",
-    "@react-navigation/material-top-tabs": "^6.6.2",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/native-stack": "^6.9.12",
     "@react-navigation/stack": "^6.3.16",
@@ -24,15 +23,11 @@
     "expo-secure-store": "~12.1.1",
     "expo-status-bar": "~1.4.4",
     "expo-updates": "~0.16.3",
-    "moment": "^2.29.4",
     "react": "18.2.0",
-    "react-hook-form": "^7.43.8",
     "react-native": "0.71.4",
     "react-native-gesture-handler": "^2.9.0",
-    "react-native-pager-view": "6.1.2",
     "react-native-safe-area-context": "4.5.0",
     "react-native-screens": "~3.20.0",
-    "react-native-tab-view": "^3.5.1",
     "sentry-expo": "~6.1.0",
     "typescript": "^4.9.4"
   },
