+++ src/components/Navigation/BottomTab/index.tsx
@@ -1,5 +1,6 @@
 import { ReactNode } from 'react';
 import { createBottomTabNavigator } from '@react-navigation/bottom-tabs';
+import CreateLogPage from 'src/pages/CreateLogPage';
 import HomePage from 'src/pages/HomePage';
 import ProfilePage from 'src/pages/ProfilePage';
 import { Icon } from 'src/components/Icon';
@@ -7,7 +8,6 @@ import SearchAndFilter from 'src/pages/HomePage/SearchAndFilter';
 import { HeaderTitle } from 'src/components/PageHeader';
 import { getBGColor, useTheme } from 'src/hooks/useTheme';
 import { Stack } from '../Stack';
-import CreateLogStack from 'src/pages/CreateLogPage';
 
 enum Tab {
   HOME = 'homeTab',
@@ -35,7 +35,6 @@ export default function BottomTabNavigation() {
           name={name}
           component={component as any} //TODO: Fix type
           options={{
-            lazy: true,
             // Page Header Options
             headerTitle: () => <HeaderTitle title={TABS_TITLE[name] ?? name} />,
             headerRight: extra,
@@ -72,7 +71,7 @@ const TABS: TabProps[] = [
   },
   {
     name: Tab.CREATE_LOG,
-    component: CreateLogStack,
+    component: CreateLogPage,
     icon: 'add-box',
     modal: ({ navigation }) => ({
       tabPress: (e) => {
